В классе LogInViewController реализована верстка экрана “Log In” согласно макету
